+++ BUILD.gn
@@ -88,6 +88,21 @@ source_set("glslang_sources") {
     "SPIRV/spvIR.h",
     "glslang/GenericCodeGen/CodeGen.cpp",
     "glslang/GenericCodeGen/Link.cpp",
+    "glslang/HLSL/hlslAttributes.cpp",
+    "glslang/HLSL/hlslAttributes.h",
+    "glslang/HLSL/hlslGrammar.cpp",
+    "glslang/HLSL/hlslGrammar.h",
+    "glslang/HLSL/hlslOpMap.cpp",
+    "glslang/HLSL/hlslOpMap.h",
+    "glslang/HLSL/hlslParseables.cpp",
+    "glslang/HLSL/hlslParseables.h",
+    "glslang/HLSL/hlslParseHelper.cpp",
+    "glslang/HLSL/hlslParseHelper.h",
+    "glslang/HLSL/hlslScanContext.cpp",
+    "glslang/HLSL/hlslScanContext.h",
+    "glslang/HLSL/hlslTokens.h",
+    "glslang/HLSL/hlslTokenStream.cpp",
+    "glslang/HLSL/hlslTokenStream.h",
     "glslang/Include/BaseTypes.h",
     "glslang/Include/Common.h",
     "glslang/Include/ConstantUnion.h",
@@ -147,21 +162,6 @@ source_set("glslang_sources") {
     "glslang/MachineIndependent/reflection.h",
     "glslang/OSDependent/osinclude.h",
     "glslang/Public/ShaderLang.h",
-    "hlsl/hlslAttributes.cpp",
-    "hlsl/hlslAttributes.h",
-    "hlsl/hlslGrammar.cpp",
-    "hlsl/hlslGrammar.h",
-    "hlsl/hlslOpMap.cpp",
-    "hlsl/hlslOpMap.h",
-    "hlsl/hlslParseHelper.cpp",
-    "hlsl/hlslParseHelper.h",
-    "hlsl/hlslParseables.cpp",
-    "hlsl/hlslParseables.h",
-    "hlsl/hlslScanContext.cpp",
-    "hlsl/hlslScanContext.h",
-    "hlsl/hlslTokenStream.cpp",
-    "hlsl/hlslTokenStream.h",
-    "hlsl/hlslTokens.h",
   ]
 
   defines = [ "ENABLE_OPT=1" ]
