+++ glslang/HLSL/hlslTokenStream.cpp
